+++ apps/api/src/app/portfolio/portfolio-calculator-baln-buy.spec.ts
@@ -1,4 +1,5 @@
 import { CurrentRateService } from '@ghostfolio/api/app/portfolio/current-rate.service';
+import { ExchangeRateDataService } from '@ghostfolio/api/services/exchange-rate-data/exchange-rate-data.service';
 import { parseDate } from '@ghostfolio/common/helper';
 import Big from 'big.js';
 
@@ -16,15 +17,24 @@ jest.mock('@ghostfolio/api/app/portfolio/current-rate.service', () => {
 
 describe('PortfolioCalculator', () => {
   let currentRateService: CurrentRateService;
+  let exchangeRateDataService: ExchangeRateDataService;
 
   beforeEach(() => {
-    currentRateService = new CurrentRateService(null, null, null);
+    currentRateService = new CurrentRateService(null, null);
+
+    exchangeRateDataService = new ExchangeRateDataService(
+      null,
+      null,
+      null,
+      null
+    );
   });
 
   describe('get current positions', () => {
     it.only('with BALN.SW buy', async () => {
       const portfolioCalculator = new PortfolioCalculator({
         currentRateService,
+        exchangeRateDataService,
         currency: 'CHF',
         orders: [
           {
@@ -63,9 +73,17 @@ describe('PortfolioCalculator', () => {
         errors: [],
         grossPerformance: new Big('24.6'),
         grossPerformancePercentage: new Big('0.09004392386530014641'),
+        grossPerformancePercentageWithCurrencyEffect: new Big(
+          '0.09004392386530014641'
+        ),
+        grossPerformanceWithCurrencyEffect: new Big('24.6'),
         hasErrors: false,
         netPerformance: new Big('23.05'),
         netPerformancePercentage: new Big('0.08437042459736456808'),
+        netPerformancePercentageWithCurrencyEffect: new Big(
+          '0.08437042459736456808'
+        ),
+        netPerformanceWithCurrencyEffect: new Big('23.05'),
         positions: [
           {
             averagePrice: new Big('136.6'),
@@ -75,17 +93,29 @@ describe('PortfolioCalculator', () => {
             firstBuyDate: '2021-11-30',
             grossPerformance: new Big('24.6'),
             grossPerformancePercentage: new Big('0.09004392386530014641'),
+            grossPerformancePercentageWithCurrencyEffect: new Big(
+              '0.09004392386530014641'
+            ),
+            grossPerformanceWithCurrencyEffect: new Big('24.6'),
             investment: new Big('273.2'),
+            investmentWithCurrencyEffect: new Big('273.2'),
             netPerformance: new Big('23.05'),
             netPerformancePercentage: new Big('0.08437042459736456808'),
+            netPerformancePercentageWithCurrencyEffect: new Big(
+              '0.08437042459736456808'
+            ),
+            netPerformanceWithCurrencyEffect: new Big('23.05'),
             marketPrice: 148.9,
+            marketPriceInBaseCurrency: 148.9,
             quantity: new Big('2'),
             symbol: 'BALN.SW',
             timeWeightedInvestment: new Big('273.2'),
+            timeWeightedInvestmentWithCurrencyEffect: new Big('273.2'),
             transactionCount: 1
           }
         ],
-        totalInvestment: new Big('273.2')
+        totalInvestment: new Big('273.2'),
+        totalInvestmentWithCurrencyEffect: new Big('273.2')
       });
 
       expect(investments).toEqual([