+++ libs/common/src/lib/interfaces/historical-data-item.interface.ts
@@ -5,11 +5,15 @@ export interface HistoricalDataItem {
   marketPrice?: number;
   netPerformance?: number;
   netPerformanceInPercentage?: number;
+  netPerformanceInPercentageWithCurrencyEffect?: number;
+  netPerformanceWithCurrencyEffect?: number;
   netWorth?: number;
   netWorthInPercentage?: number;
   quantity?: number;
   totalAccountBalance?: number;
   totalInvestment?: number;
+  totalInvestmentValueWithCurrencyEffect?: number;
   value?: number;
   valueInPercentage?: number;
+  valueWithCurrencyEffect?: number;
 }