+++ CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Broken down the performance into asset and currency on the analysis page (experimental)
+
 ### Changed
 
+- Prepared the portfolio calculation for exchange rate effects
 - Upgraded `prettier` from version `3.1.1` to `3.2.1`
 
 ## 2.37.0 - 2024-01-11